Move bounds cache to showWindow method
This allows showWindow to be called manually and the window will appear in the correct place assuming the tray has already been clicked at some stage.

My use case is that I want the tray to open on the `drop-files` event. Previous to this, calling `showWindow(null)` would position the window at the top-right (OS X). This is not an ideal solution because if the tray hasn't been clicked previously then the window will still appear in the top-right. A better solution could be implemented upstream in Electron (see: https://github.com/atom/electron/issues/3377).

This is a breaking change because it changes the behaviour of `menuBar.showWindow(null)` after the tray has been clicked.

This may not be a desirable PR for the reasons mentioned above but I'll submit it for reference anyway.